requirements.txt
# LangChain and LangGraph Core Libraries
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.0
langchain-classic>=0.1.0
langchain-experimental>=0.1.0

# Vector Database
chromadb>=0.4.0

# PDF Processing
pymupdf>=1.23.0  # fitz
pdfplumber>=0.10.0
pytesseract>=0.3.10
pdf2image>=1.16.0
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0

# Optional but recommended
langdetect>=1.0.9
langchain-cohere>=0.1.0
